# api-contracts/api-contract.yml
swagger: '2.0'
info:
  version: 0.0.1
  title: Capability Service
  description: The authority of all things CAPABILITY related!
consumes:
  - application/json
produces:
  - application/json
schemes: 
  - http
host: localhost:5000
basePath: /api/v1/
paths:
  /capabilities: 
    get:
      summary: Returns a list of capabilities
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/capability-list'
    post:
      summary: Create a new capability
      parameters: 
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              name:
                type: string
      responses:
        201:
          description: Created
          headers:
            location:
              type: string
              description: URI where the capability can be retrieved
          schema:
            $ref: '#/definitions/capability'
        400:
          description: BadRequest
          schema:
            $ref: '#/definitions/errorObject'
  /capabilities/{capabilityId}:
    get:
      summary: Returns a single capability
      parameters:
        - in: path
          name: capabilityId
          type: string
          required: true
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/capability'
  /capabilities/{capabilityId}/members:
    post:
      summary: Adds a person to a capability 
      parameters:
        - in: path
          name: capabilityId
          required: true
          type: string
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              email:
                type: string
      responses:
        200:
          description: Ok
        404:
          description: Not Found
          schema:
            $ref: '#/definitions/errorObject'
  /capabilities/{capabilityId}/members/{email}:
    delete:
      summary: Deletes a person from a capability 
      parameters:
        - in: path
          name: capabilityId
          required: true
          type: string
        - in: path
          name: email
          type: string
          required: true
      responses:
        200:
          description: Ok
        404:
          description: Not Found
  /capabilities/{capabilityId}/contexts:
    post:
      summary: Adds a context to a capability 
      parameters:
        - in: path
          name: capabilityId
          required: true
          type: string
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              name:
                type: string
      responses:
        200:
          description: Ok
        404:
          description: Not Found
          schema:
            $ref: '#/definitions/errorObject'
definitions:
  capability-list:
    type: object
    properties:
      items:
        type: array
        items:
          $ref: '#/definitions/capability'
  capability:
    type: object
    properties:
      id:
        type: string
      name:
        type: string
      description:
        type: string
      members:
        type: array
        items:
          $ref: '#/definitions/member'
      contexts:
        type: array
        items:
          $ref: '#/definitions/context'
  member:
    type: object
    properties:
      email:
        type: string
  context:
    type: object
    properties:
      id:
        type: string
      name:
        type: string
  errorObject:
    type: object
    properties:
      message:
        type: string